requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
colorama==0.4.5
colorlog==6.7.0
frozenlist==1.3.1
hikari==2.0.0.dev110
hikari-lightbulb==2.2.4
idna==3.4
multidict==6.0.2
yarl==1.8.1
brainfuckery==1.0.1
